now it's actually caught with asan
